requirements.txt
ldap3==2.9.1
pyasn1==0.4.8
python-dotenv==0.21.1
dataclasses-json==0.5.7
typing-inspect==0.8.0
typing_extensions==4.5.0
mypy-extensions==1.0.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
